+++ src/main/java/org/jabref/gui/shared/SharedDatabaseUIManager.java
@@ -175,6 +175,7 @@ public class SharedDatabaseUIManager {
                 bibDatabaseContext,
                 tabContainer,
                 dialogService,
+                aiService,
                 preferences,
                 stateManager,
                 fileUpdateMonitor,